requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.6
Werkzeug==2.2.2
